third_party/blink/web_tests/wpt_internal/compression-dictionary/supported-rel-list.https.html

<!DOCTYPE html>
<title>Supported rel list for dictionary</title>

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/compression-dictionary-origin-trial.js"></script>

<body>
  <script>
    test(() => {
      assert_true(document.createElement('link').relList.supports('dictionary'),
                  "dictionary links shall be supported by default.");
    }, 'rel=dictionary correctly supported');

    test(() => {
      enableCompressionDictionaryOriginTrial();
      assert_true(document.createElement('link').relList.supports('dictionary'),
                  "dictionary links shall be supported with origin trial.");
    }, 'rel=dictionary correctly supported with origin trail tokens');
  </script>
</body>
